requirements.txt
requests
discord
python-dotenv
waitress
Flask
